python-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
pandas==2.1.4
numpy==1.24.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
pymongo==4.6.0
redis==5.0.1
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3.post1 